--- helloworld/pom.xml
@@ -1,143 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>alexa-skills-kit-samples</groupId>
-  <artifactId>blackjack</artifactId>
-  <packaging>jar</packaging>
-  <version>1.0</version>
-  <name>blackjack</name>
-  <url>http://developer.amazon.com/ask</url>
-
-  <repositories>
-    <repository>
-      <id>alexa-skills-kit-repo</id>
-      <url>file://${project.basedir}/repo</url>
-    </repository>
-  </repositories>
-
-  <dependencies>
-    <dependency>
-      <groupId>com.amazon.alexa</groupId>
-      <artifactId>alexa-skills-kit</artifactId>
-      <version>1.8.1</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.jetty</groupId>
-      <artifactId>jetty-server</artifactId>
-      <version>9.0.6.v20130930</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.eclipse.jetty</groupId>
-      <artifactId>jetty-servlet</artifactId>
-      <version>9.0.6.v20130930</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>log4j</groupId>
-      <artifactId>log4j</artifactId>
-      <version>1.2.17</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.slf4j</groupId>
-      <artifactId>slf4j-api</artifactId>
-      <version>1.7.10</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.slf4j</groupId>
-      <artifactId>slf4j-log4j12</artifactId>
-      <version>1.7.10</version>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.commons</groupId>
-      <artifactId>commons-lang3</artifactId>
-      <version>3.4</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.directory.studio</groupId>
-      <artifactId>org.apache.commons.io</artifactId>
-      <version>2.4</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.amazonaws</groupId>
-      <artifactId>aws-lambda-java-core</artifactId>
-      <version>1.0.0</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.amazonaws</groupId>
-      <artifactId>aws-lambda-java-log4j</artifactId>
-      <version>1.0.0</version>
-      <scope>compile</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.amazonaws</groupId>
-      <artifactId>aws-java-sdk-dynamodb</artifactId>
-      <version>1.9.40</version>
-    </dependency>
-  </dependencies>
-
-  <properties>
-    <property name="disableRequestSignatureCheck" value="false"/>
-    <property name="supportedApplicationIds" value=""/>
-    <property name="timestampTolerance" value="150"/>
-  </properties>
-
-  <build>
-    <sourceDirectory>src</sourceDirectory>
-    <resources>
-      <resource>
-        <directory>src/resources</directory>
-      </resource>
-    </resources>
-    <pluginManagement>
-      <plugins>
-        <plugin>
-          <groupId>org.apache.maven.plugins</groupId>
-          <artifactId>maven-compiler-plugin</artifactId>
-          <version>3.3</version>
-          <configuration>
-            <source>1.7</source>
-            <target>1.7</target>
-          </configuration>
-        </plugin>
-
-        <plugin>
-          <groupId>org.codehaus.mojo</groupId>
-          <artifactId>exec-maven-plugin</artifactId>
-          <version>1.2.1</version>
-          <configuration>
-            <mainClass>Launcher</mainClass>
-            <systemProperties>
-              <systemProperty>
-                <key>javax.net.ssl.keyStore</key>
-                <value>/insert/your/path/java-keystore.jks</value>
-              </systemProperty>
-              <systemProperty>
-                <key>javax.net.ssl.keyStorePassword</key>
-                <value>insert_your_password</value>
-              </systemProperty>
-              <systemProperty>
-                <key>com.amazon.speech.speechlet.servlet.disableRequestSignatureCheck</key>
-                <value>${disableRequestSignatureCheck}</value>
-              </systemProperty>
-              <systemProperty>
-                <key>com.amazon.speech.speechlet.servlet.supportedApplicationIds</key>
-                <value>${supportedApplicationIds}</value>
-              </systemProperty>
-              <systemProperty>
-                <key>com.amazon.speech.speechlet.servlet.timestampTolerance</key>
-                <value>${timestampTolerance}</value>
-              </systemProperty>
-            </systemProperties>
-          </configuration>
-        </plugin>
-      </plugins>
-    </pluginManagement>
-  </build>
-</project>
