--- helloworld/speechAssets/IntentSchema.json
@@ -1,10 +1,0 @@
-{
-  "intents": [
-    {
-      "intent": "HelloWorldIntent"
-    },
-    {
-      "intent": "AMAZON.HelpIntent"
-    }
-  ]
-}
